# vlife-admin/src/main/resources/static/doc/openapi.json
{
  "components": {
    "schemas": {
      "ProjectDto": {
        "properties": {
          "id": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "accountId": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          }
        }
      },
...
      "ProjectStep": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "projectId": {
            "format": "string",
            "description": "项目",
            "type": "string"
          },
          "stepId": {
            "format": "string",
            "description": "项目的阶段",
            "type": "string"
          }
        }
      },
      "Task": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          },
          "projectId": {
            "format": "string",
            "description": "所属项目",
            "type": "string"
          },
          "point": {
            "format": "int32",
            "description": "任务点数",
            "type": "integer"
          },
          "progress": {
            "format": "double",
            "description": "当前完成进度",
            "type": "number"
          },
          "state": {
            "format": "string",
            "description": "任务完成状态",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "任务名称",
            "type": "string"
          },
          "sort": {
            "format": "int32",
            "description": "排序",
            "type": "integer"
          },
          "level": {
            "format": "string",
            "description": "紧急程度(字典)",
            "type": "string"
          },
          "stepId": {
            "format": "string",
            "description": "当前所在阶段",
            "type": "string"
          },
          "accountId": {
            "format": "string",
            "description": "任务创建人",
            "type": "string"
          }
        }
      },
...
      "Step": {
        "properties": {
          "id": {
            "format": "string",
            "description": "主键id",
            "type": "string"
          },
          "status": {
            "format": "string",
            "description": "实体状态 【1有效 0逻辑删除】",
            "type": "string"
          },
          "createDate": {
            "format": "string",
            "description": "记录创建日期",
            "type": "string"
          },
          "modifyDate": {
            "format": "string",
            "description": "最近修改日期",
            "type": "string"
          }
        }
      },
...
      "ProjectDetailVo": {
        "properties": {
          "id": {
            "format": "string",
            "description": "No comments found.",
            "type": "string"
          },
          "name": {
            "format": "string",
            "description": "项目名称",
            "type": "string"
          },
          "account_name": {
            "format": "string",
            "description": "项目负责人姓名",
            "type": "string"
          },
          "planDate": {
            "format": "string",
            "description": "计划开始日期",
            "type": "string"
          },
          "finishedDate": {
            "format": "string",
            "description": "结束日期",
            "type": "string"
          },
          "state": {
            "format": "string",
            "description": "项目状态 project_state",
            "type": "string"
          },
          "type": {
            "format": "string",
            "description": "项目类型",
            "type": "string"
          },
          "totalPoint": {
            "format": "int32",
            "description": "项目任务总点数",
            "type": "integer"
          },
          "finishedPoint": {
            "format": "int32",
            "description": "已经完成的任务点数",
            "type": "integer"
          },
          "steps": {
            "description": "项目的生命周期",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Step"
            }
          },
          "tasks": {
            "description": "项目下的任务",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Task"
            }
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://localhost:8088"
    }
  ],
  "openapi": "3.0.3",
  "paths": {
    "/projectStep/detail/{id}": {
      "get": {
        "summary": "明细查询项目阶段关系表;",
        "deprecated": false,
        "description": "明细查询项目阶段关系表;",
        "tags": [
          "项目阶段关系表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStep"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"goe233\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"stepId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
    "/task/list": {
      "get": {
        "summary": "列表查询任务表;",
        "deprecated": false,
        "description": "列表查询任务表;",
        "tags": [
          "任务表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "[\n  {\n    \"id\": \"144\",\n    \"status\": \"vdt985\",\n    \"createDate\": \"2022-08-13\",\n    \"modifyDate\": \"2022-08-13\",\n    \"projectId\": \"144\",\n    \"point\": 830,\n    \"progress\": 28.83,\n    \"state\": \"0ns1bd\",\n    \"name\": \"ellie.kuphal\",\n    \"sort\": 636,\n    \"level\": \"7vcp83\",\n    \"stepId\": \"144\",\n    \"accountId\": \"144\"\n  }\n]"
                  }
                }
              }
            }
          }
        },
        "operationId": "list",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "项目id",
            "required": false,
            "example": "144",
            "name": "projectId"
          },
          {
            "schema": {
              "type": "array",
              "items": {}
            },
            "in": "query",
            "description": "进度范围",
            "required": false,
            "example": "94.38,97.97",
            "name": "progress"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "任务名称",
            "required": false,
            "example": "ellie.kuphal",
            "name": "name"
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "description": "任务状态",
            "required": false,
            "example": "543kga",
            "name": "state"
          }
        ]
      }
    },
...
    "/step/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "项目阶段接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "411"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/projectStep/save": {
      "post": {
        "summary": "保存项目阶段关系表;",
        "deprecated": false,
        "description": "保存项目阶段关系表;",
        "tags": [
          "项目阶段关系表接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectStep"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"status\": \"lcc6s9\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"stepId\": \"144\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectStep"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"y5f0dn\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"stepId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
    "/project/detail/{id}": {
      "get": {
        "summary": "明细查询类说明;",
        "deprecated": false,
        "description": "明细查询类说明;",
        "tags": [
          "项目表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDetailVo"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"name\": \"ellie.kuphal\",\n  \"account_name\": \"ellie.kuphal\",\n  \"planDate\": \"2022-08-13\",\n  \"finishedDate\": \"2022-08-13\",\n  \"state\": \"ai7vkf\",\n  \"type\": \"uxhpr2\",\n  \"totalPoint\": 244,\n  \"finishedPoint\": 526,\n  \"steps\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"n71t4f\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\"\n    }\n  ],\n  \"tasks\": [\n    {\n      \"id\": \"144\",\n      \"status\": \"bjtpye\",\n      \"createDate\": \"2022-08-13\",\n      \"modifyDate\": \"2022-08-13\",\n      \"projectId\": \"144\",\n      \"point\": 715,\n      \"progress\": 44.70,\n      \"state\": \"3l9a3j\",\n      \"name\": \"ellie.kuphal\",\n      \"sort\": 418,\n      \"level\": \"rp0mi7\",\n      \"stepId\": \"144\",\n      \"accountId\": \"144\"\n    }\n  ]\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/step/save": {
      "post": {
        "summary": "保存项目阶段;",
        "deprecated": false,
        "description": "保存项目阶段;",
        "tags": [
          "项目阶段接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Step"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"status\": \"pnoey6\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Step"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"3yjvir\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/project/save": {
      "post": {
        "summary": "保存项目编辑;",
        "deprecated": false,
        "description": "保存项目编辑;",
        "tags": [
          "项目表接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectDto"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"name\": \"ellie.kuphal\",\n  \"accountId\": \"144\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectDto"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"name\": \"ellie.kuphal\",\n  \"accountId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/step/detail/{id}": {
      "get": {
        "summary": "明细查询项目阶段;",
        "deprecated": false,
        "description": "明细查询项目阶段;",
        "tags": [
          "项目阶段接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Step"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"uug4k2\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/task/detail/{id}": {
      "get": {
        "summary": "明细查询任务表;",
        "deprecated": false,
        "description": "明细查询任务表;",
        "tags": [
          "任务表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"929aou\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"point\": 785,\n  \"progress\": 12.27,\n  \"state\": \"d70abn\",\n  \"name\": \"ellie.kuphal\",\n  \"sort\": 227,\n  \"level\": \"f55he8\",\n  \"stepId\": \"144\",\n  \"accountId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "detail",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
    "/task/save": {
      "post": {
        "summary": "保存任务表;",
        "deprecated": false,
        "description": "保存任务表;",
        "tags": [
          "任务表接口;"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Task"
              },
              "examples": {
                "json": {
                  "summary": "test data",
                  "value": "{\n  \"id\": \"144\",\n  \"status\": \"y5tvtb\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"point\": 679,\n  \"progress\": 78.69,\n  \"state\": \"s5gsmc\",\n  \"name\": \"ellie.kuphal\",\n  \"sort\": 40,\n  \"level\": \"apxwq0\",\n  \"stepId\": \"144\",\n  \"accountId\": \"144\"\n}"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Task"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "{\n  \"id\": \"144\",\n  \"status\": \"49j9gg\",\n  \"createDate\": \"2022-08-13\",\n  \"modifyDate\": \"2022-08-13\",\n  \"projectId\": \"144\",\n  \"point\": 42,\n  \"progress\": 23.34,\n  \"state\": \"qtspsm\",\n  \"name\": \"ellie.kuphal\",\n  \"sort\": 904,\n  \"level\": \"wczmw1\",\n  \"stepId\": \"144\",\n  \"accountId\": \"144\"\n}"
                  }
                }
              }
            }
          }
        },
        "operationId": "save",
        "parameters": []
      }
    },
...
    "/projectStep/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "项目阶段关系表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "263"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
    "/task/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "任务表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "146"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
    "/project/remove/{id}": {
      "delete": {
        "summary": "逻辑删除;",
        "deprecated": false,
        "description": "逻辑删除;",
        "tags": [
          "项目表接口;"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "number"
                },
                "examples": {
                  "json": {
                    "summary": "test data",
                    "value": "275"
                  }
                }
              }
            }
          }
        },
        "operationId": "remove",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "description": "主键id;",
            "required": true,
            "example": "144",
            "name": "id"
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "test",
    "version": "1.0.0"
  },
...
}